Add Parsec.evalValues(list) to eval in batch and upgrade project to AndroidX (#18)
This PR aims to implement multiple equation calculus. 

Now there is a new Parsec.evalValues(list<string>) function that accepts a list of equations, calculates each one, and returns a list of values. If any equation in the list is incorrect, the result of this equation will be `null` and the rest of the calculations will not be affected.

This PR also updates the `gradle` and `kotlin` versions and implements `androidx` in the demo module.

### Sample:

![image](https://user-images.githubusercontent.com/13650290/142571509-31ccc171-6dfa-4f1a-b049-ceb026b62b02.png)