Update version of Gradle and libraries (#29)
## Description

This PR aims to upgrade the gradle version, and also the libraries used in the project. The table below shows the upgrades:

| Artifact | Old version | New version |
| --------- | -------------- | ---------------- |
| Gradle |  `6.7.1` | `7.5` |
| Kotlin |  `1.5.31` | `1.9.10` |
| Android SDK |  `30` | `33` |
| android-gradle-plugin |  `4.2.2` | `7.4.2` |
| appcompat |  `1.3.1` | `1.6.1` |
| constraintlayout |  `2.1.2` | `2.1.4` |
| junit |  `1.1.3` | `1.1.5` |
| espresso-core |  `3.4.0` | `3.5.1` |

### View Binding changes

In this new version of the Android SDK, Kotlin Android Extensions are deprecated, which means that using Kotlin synthetics for view binding is no longer supported. How the project used Kotlin synthetics for view binding, this [guide](https://developer.android.com/topic/libraries/view-binding/migration) was used to migrate to Jetpack view binding.

For this it was necessary: 

- Set the `viewBinding` build option to `true` in the module-level `build.gradle` file:

```
android {
    ...
    buildFeatures {
        viewBinding = true
    }
}
```
- Inflate the generated binding class instance for the activity to use (guide [here](https://developer.android.com/topic/libraries/view-binding#activities)):

To set up an instance of the binding class for use with an activity, perform the following steps in the activity's `onCreate() `method:

1. Call the static inflate() method included in the generated binding class. This creates an instance of the binding class for the activity to use.
2. Get a reference to the root view by either calling the `getRoot()` method or using Kotlin property syntax.
3. Pass the root view to `setContentView()` to make it the active view on the screen.

These steps resulted in the following code:

```
...
private lateinit var binding: ActivityMainBinding

override fun onCreate(savedInstanceState: Bundle?) {
        super.onCreate(savedInstanceState)
        binding = ActivityMainBinding.inflate(layoutInflater)
        setContentView(binding.root)
}
...
```